pyproject.toml
[project]
name = "hotel-recommendations"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "agency-swarm>=0.7.2",
    "fastapi>=0.116.1",
    "fastmcp>=2.11.2",
    "google-generativeai>=0.8.5",
    "googlemaps>=4.10.0",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "orjson>=3.11.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "tenacity>=9.1.2",
    "uvicorn>=0.35.0",
]
